Export reconciler Validate method
Export the Validate() method for Reconcilers and SubReconcilers. Test
cases detect this method and call it before the test is run to fail fast
with a meaningful error for ambiguous reconciler definitions.

Since the reconciler validation tests can now live in a _test package,
they moved to live next to the primary tests for their respective
reconciler.

Tests previously using invalid reconcilers have been corrected or
removed.
