+++ reconcilers/cast_test.go
@@ -33,6 +33,7 @@ import (
 	"reconciler.io/runtime/internal/resources/dies"
 	"reconciler.io/runtime/reconcilers"
 	rtesting "reconciler.io/runtime/testing"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func TestCastResource(t *testing.T) {
@@ -239,4 +240,50 @@ func TestCastResource(t *testing.T) {
 	rts.Run(t, scheme, func(t *testing.T, rtc *rtesting.SubReconcilerTestCase[*resources.TestResource], c reconcilers.Config) reconcilers.SubReconciler[*resources.TestResource] {
 		return rtc.Metadata["SubReconciler"].(func(*testing.T, reconcilers.Config) reconcilers.SubReconciler[*resources.TestResource])(t, c)
 	})
+}
+
+func TestCastResource_Validate(t *testing.T) {
+	tests := []struct {
+		name       string
+		resource   client.Object
+		reconciler *reconcilers.CastResource[*corev1.ConfigMap, *corev1.Secret]
+		shouldErr  string
+	}{
+		{
+			name:       "empty",
+			resource:   &corev1.ConfigMap{},
+			reconciler: &reconcilers.CastResource[*corev1.ConfigMap, *corev1.Secret]{},
+			shouldErr:  `CastResource "" must define Reconciler`,
+		},
+		{
+			name:     "valid",
+			resource: &corev1.ConfigMap{},
+			reconciler: &reconcilers.CastResource[*corev1.ConfigMap, *corev1.Secret]{
+				Reconciler: &reconcilers.SyncReconciler[*corev1.Secret]{
+					Sync: func(ctx context.Context, resource *corev1.Secret) error {
+						return nil
+					},
+				},
+			},
+		},
+		{
+			name:     "missing reconciler",
+			resource: &corev1.ConfigMap{},
+			reconciler: &reconcilers.CastResource[*corev1.ConfigMap, *corev1.Secret]{
+				Name:       "missing reconciler",
+				Reconciler: nil,
+			},
+			shouldErr: `CastResource "missing reconciler" must define Reconciler`,
+		},
+	}
+
+	for _, c := range tests {
+		t.Run(c.name, func(t *testing.T) {
+			ctx := reconcilers.StashResourceType(context.TODO(), c.resource)
+			err := c.reconciler.Validate(ctx)
+			if (err != nil) != (c.shouldErr != "") || (c.shouldErr != "" && c.shouldErr != err.Error()) {
+				t.Errorf("validate() error = %q, shouldErr %q", err, c.shouldErr)
+			}
+		})
+	}
 }
